refactor(fluid): 重构泵块相关类的包结构并优化客户端渲染逻辑
- 将 PumpBlock 从 dev.dubhe.anvilcraft.block 移至 dev.dubhe.anvilcraft.block.fluid
- 将 PumpBlockEntity 从 dev.dubhe.anvilcraft.block.entity 移至 dev.dubhe.anvilcraft.block.entity.fluid
- 更新 AbstractPipeBlockEntity、PipeBlockEntity、PipeNodeBlockEntity 中的导入路径
- 更新 ModBlocks 和 ModBlockEntities 中的导入路径以匹配新的包结构
- 在 PipeBlockEntity 中提前声明 effectiveHeight 变量避免重复定义
- 优化 PumpBlockEntity 的客户端渲染逻辑，直接设置 working 状态
