+++ src/main/java/dev/dubhe/anvilcraft/block/entity/fluid/AbstractPipeBlockEntity.java
@@ -1,7 +1,6 @@
 package dev.dubhe.anvilcraft.block.entity.fluid;
 
-import dev.dubhe.anvilcraft.block.PumpBlock;
-import dev.dubhe.anvilcraft.block.entity.PumpBlockEntity;
+import dev.dubhe.anvilcraft.block.fluid.PumpBlock;
 import dev.dubhe.anvilcraft.block.fluid.PipeBlock;
 import dev.dubhe.anvilcraft.block.fluid.PipeCornerBlock;
 import dev.dubhe.anvilcraft.block.fluid.PipeNodeBlock;
